bruh moment

+++ packets/common/src/main/java/org/screamingsandals/lib/packet/PacketWriter.java
@@ -43,7 +43,7 @@ public abstract class PacketWriter extends OutputStream {
         buffer.writeChar(c);
     }
 
-    public void writeShort(short s) {
+    public void writeShort(int s) {
         buffer.writeShort(s);
     }
 
@@ -205,15 +205,15 @@ public abstract class PacketWriter extends OutputStream {
     }
 
     public void writeMotion(Vector3D vector3D) {
-        writeShort((short) (vector3D.getX() * 8000));
-        writeShort((short) (vector3D.getY() * 8000));
-        writeShort((short) (vector3D.getZ() * 8000));
+        writeShort((int) (vector3D.getX() * 8000));
+        writeShort((int) (vector3D.getY() * 8000));
+        writeShort((int) (vector3D.getZ() * 8000));
     }
 
     public void writeMove(Vector3D vector3D) {
-        writeShort((short) (vector3D.getX() * 4096));
-        writeShort((short) (vector3D.getY() * 4096));
-        writeShort((short) (vector3D.getZ() * 4096));
+        writeShort((int) (vector3D.getX() * 4096));
+        writeShort((int) (vector3D.getY() * 4096));
+        writeShort((int) (vector3D.getZ() * 4096));
     }
 
     public void writeItem(Item item) {
@@ -221,7 +221,7 @@ public abstract class PacketWriter extends OutputStream {
             if (protocol() >= 402) {
                 writeBoolean(false);
             } else {
-                writeShort((short) -1);
+                writeShort(-1);
             }
         } else {
             if (protocol() >= 402) {
@@ -229,12 +229,12 @@ public abstract class PacketWriter extends OutputStream {
 
                 writeVarInt(getItemId(item.getMaterial()));
             } else {
-                writeShort((short) getItemId(item.getMaterial()));
+                writeShort(getItemId(item.getMaterial()));
             }
 
             write(item.getAmount());
             if (protocol() < 351) {
-                writeShort((short) item.getMaterial().getDurability());
+                writeShort(item.getMaterial().getDurability());
             }
             writeNBTFromItem(item);
         }
